requirements.txt
numpy<2.0
llama-cpp-python
requests
